update: Grid.h. size_t -> int.

+++ include/Grid.h
@@ -9,12 +9,12 @@ class Grid {
 public:
     Grid() = default;
 
-    static constexpr size_t rows() { return rows_; }
-    static constexpr size_t cols() { return cols_; }
-    static constexpr size_t tiles() { return rows_ * cols_; };
+    static constexpr int rows() { return rows_; }
+    static constexpr int cols() { return cols_; }
+    static constexpr int tiles() { return rows_ * cols_; };
 
     ftxui::Canvas toCanvas(ftxui::Canvas& cva) const;
 private:
-    inline static constexpr size_t rows_{10};
-    inline static constexpr size_t cols_{10};
+    inline static constexpr int rows_{10};
+    inline static constexpr int cols_{10};
 };
